[pf-579] Rename SmartRandom to dash-case
This file was renamed in a previous commit. Unfortunately, the
previous commit has already been pushed and merged. It also deletes
a temp file that happened to have been commited by mistake
